+++ src/testMethod.js
@@ -14,6 +14,16 @@
  */
 export function test(pattern, str){
   if (arguments.length === 1) return _str => test(pattern, _str)
+  console.log({
+    str,
+    pattern,
+  })
+
+  if (typeof pattern === 'string'){
+    throw new TypeError(
+      `‘test’ requires a value of type RegExp as its first argument; received "${ pattern }"`
+    )
+  }
 
   return str.search(pattern) !== -1
 }