+++ src/testMethod.spec.js
@@ -5,3 +5,9 @@ test('', () => {
 
   expect(testMethod(/^y/)('xyz')).toBeFalsy()
 })
+
+test('throws if first argument is not regex', () => {
+  expect(
+    () => testMethod('foo', 'bar')
+  ).toThrow('1')
+})