+++ src/range.js
@@ -15,6 +15,12 @@
 export function range(from, to){
   if (arguments.length === 1) return _to => range(from, _to)
 
+  if (Number.isNaN(Number(from)) || Number.isNaN(Number(to))){
+    throw new TypeError('Both arguments to range must be numbers')
+  }
+
+  if (to < from) return []
+
   const len = to - from
   const willReturn = Array(len)