+++ src/range.spec.js
@@ -1,9 +1,21 @@
 import { range } from './range'
 
-test('1', () => {
+test('happy', () => {
   expect(range(0, 10)).toEqual([ 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 ])
 })
 
-test('2', () => {
+test('end range is bigger than start range', () => {
+  expect(range(7, 3)).toEqual([])
+  expect(range(5, 5)).toEqual([])
+})
+
+test('with bad input', () => {
+  expect(
+    () => range('a', 'z')
+  ).toThrow('Both arguments to range must be numbers')
+
+})
+
+test('curry', () => {
   expect(range(0)(10)).toEqual([ 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 ])
 })