+++ src/internal/tryOuts.spec.js
@@ -1,0 +1,12 @@
+function isNumberA(x){
+  return Number.isInteger(x) && x > 0
+}
+
+function isNumberB(x){
+  return !Number.isNaN(Number(x))
+}
+
+function isNumberC(x){
+  return !Number.isNaN(Number(x))
+}
+