+++ src/range.spec.js
@@ -11,7 +11,10 @@ test('end range is bigger than start range', () => {
 
 test('with bad input', () => {
   expect(
-    () => range('a', 'z')
+    () => range('a', 6)
+  ).toThrow('Both arguments to range must be numbers')
+  expect(
+    () => range(6, 'z')
   ).toThrow('Both arguments to range must be numbers')
 
 })