+++ src/testMethod.js
@@ -14,10 +14,6 @@
  */
 export function test(pattern, str){
   if (arguments.length === 1) return _str => test(pattern, _str)
-  console.log({
-    str,
-    pattern,
-  })
 
   if (typeof pattern === 'string'){
     throw new TypeError(