+++ src/times.js
@@ -1,4 +1,4 @@
-import { map } from './map'
+import { forEach } from './forEach'
 import { range } from './range'
 
 /**
@@ -20,6 +20,7 @@ import { range } from './range'
  */
 export function times(fn, n){
   if (arguments.length === 1) return _n => times(fn, _n)
+  if (!Number.isInteger(n) || n < 0) throw new RangeError('n must be an integer')
 
-  return map(fn, range(0, n))
+  return forEach(fn, range(0, n))
 }