+++ src/propEq.spec.js
@@ -7,3 +7,8 @@ test('propEq', () => {
 
   expect(propEq('foo')('bar')({ foo : 'baz' })).toBeFalsy()
 })
+
+test('happy', () => {
+  expect(propEq('name', 'Abby', null)).toEqual(false)
+// expect(propEq('name', 'Abby', undefined)).toEqual(false)
+})