+++ src/testMethod.spec.js
@@ -10,6 +10,6 @@ test('throws if first argument is not regex', () => {
   expect(
     () => testMethod('foo', 'bar')
   ).toThrow(
-    `‘test’ requires a value of type RegExp as its first argument; received "foo"`
+    '‘test’ requires a value of type RegExp as its first argument; received "foo"'
   )
 })