+++ src/times.js
@@ -1,4 +1,4 @@
-import { forEach } from './forEach'
+import { map } from './map'
 import { range } from './range'
 
 /**
@@ -22,5 +22,5 @@ export function times(fn, n){
   if (arguments.length === 1) return _n => times(fn, _n)
   if (!Number.isInteger(n) || n < 0) throw new RangeError('n must be an integer')
 
-  return forEach(fn, range(0, n))
+  return map(fn, range(0, n))
 }