+++ src/pipe.js
@@ -17,5 +17,6 @@ import { compose } from './compose'
  * @symb R.pipe(f, g, h)(a, b) = h(g(f(a, b)))
  */
 export function pipe(...fns){
+  if(fns.length === 0) throw new Error('pipe requires at least one argument')
   return compose(...fns.reverse())
 }