+++ CHANGELOG.md
@@ -40,6 +40,8 @@ All breaking changes:
 
 -- R.range work with bad inputs
 
+-- R.times work with bad inputs
+
 -- R.reverse works with strings
 
 -- R.splitEvery throws on non-positive integer index